% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_gsea.R
\name{iterate_gsea}
\alias{iterate_gsea}
\title{Iterate GSEA}
\usage{
iterate_gsea(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  x_quantiles = 10,
  y_quantiles = 10,
  use_quantiles = 10,
  nCores = 1
)
}
\arguments{
\item{xmat}{gene x cell type matrix.}

\item{ymat}{gene x trait matrix.}

\item{correction_method}{Multiple-testing correction method 
to be passed to \code{stats::p.adjust}.}

\item{qvalue_thresh}{q.value threshold to use when report 
significant results summary.}

\item{x_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{y_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{use_quantiles}{Which quantiles in to use in 
\code{xmat} and \code{ymat}.}

\item{nCores}{Number of cores to use in parallel. 
Will optimize if \code{NULL}.}
}
\value{
\code{data.table} of enrichment results.
}
\description{
Iterate Gene Set Enrichment Analysis (GSEA) across all traits and cell types.
}
\examples{
set.seed(42)

### x matrix (e.g. cell type expression specificity signatures)
xmat <- simulate_matrix(nrow = 200, ncol = 10, col_prefix="celltype_")

### y matrix (e.g. gene-phenotype association signatures)
ymat <- simulate_matrix(nrow = 100, ncol = 50, col_prefix="phenotype_") 

res_gsea <- iterate_gsea(xmat = xmat,
                         ymat = ymat,
                         nCores = 1)
}
