% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize_matrix.R
\name{quantize_matrix}
\alias{quantize_matrix}
\title{Quantize matrix}
\usage{
quantize_matrix(X, n = 40, as_sparse = TRUE, verbose = TRUE)
}
\arguments{
\item{X}{Input matrix.}

\item{n}{Number of quantiles.}

\item{as_sparse}{Convert the matrix to sparse format.}

\item{verbose}{Print messages.}
}
\description{
Quantize matrix
}
\examples{
set.seed(42)
X <-simulate_matrix(nrow=100, ncol=10)
Xq <- quantize_matrix(X)
}
