% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize_vector.R
\name{quantize_vector}
\alias{quantize_vector}
\title{Quantize vector}
\usage{
quantize_vector(vec, n = 40, default = as.integer(n/2))
}
\arguments{
\item{vec}{A vector.}

\item{n}{Number of quantiles.}

\item{default}{Default value to assign to values that fall outside
the range of quantiles.}
}
\value{
A vector of quantiles.
}
\description{
Transform a vector into n quantiles.
}
\examples{
vec <- runif(100)
qvec <- quantize_vector(vec)
}
