% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(
  mat,
  transpose = TRUE,
  center = TRUE,
  scale. = FALSE,
  rank. = NULL,
  ...
)
}
\arguments{
\item{mat}{Matrix to run PCA on.}

\item{transpose}{Whether to transpose the matrix first.}

\item{center}{a logical value indicating whether the variables
    should be shifted to be zero centered. Alternately, a vector of
    length equal the number of columns of \code{x} can be supplied.
    The value is passed to \code{scale}.}

\item{scale.}{a logical value indicating whether the variables should
    be scaled to have unit variance before the analysis takes
    place.  The default is \code{FALSE} for consistency with S, but
    in general scaling is advisable.  Alternatively, a vector of length
    equal the number of columns of \code{x} can be supplied.  The
    value is passed to \code{\link{scale}}.}

\item{rank.}{optionally, a number specifying the maximal rank, i.e.,
    maximal number of principal components to be used.  Can be set as
    alternative or in addition to \code{tol}, useful notably when the
    desired rank is considerably smaller than the dimensions of the matrix.}

\item{...}{Additional parameters passed to \link[stats]{prcomp}.}
}
\description{
Run Principal Components Analysis (PCA).
}
\details{
Uses \link[stats]{prcomp}.
}
