% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cores.R
\name{set_cores}
\alias{set_cores}
\title{Set cores}
\usage{
set_cores(workers = 0.9, progressbar = TRUE, verbose = TRUE)
}
\arguments{
\item{workers}{Number (>1) or proportion (<1) of worker cores to use.}

\item{progressbar}{logical(1) Enable progress bar
(based on \code{plyr:::progress_text}).
 Enabling the progress bar changes the default value of tasks to
 \code{.Machine$integer.max}, so that progress is reported for
 each element of X.}

\item{verbose}{Print messages.}
}
\value{
List of core allocations.
}
\description{
Assign cores automatically for parallel processing, while reserving some.
}
\keyword{internal}
