% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_gsea.R
\name{iterate_gsea}
\alias{iterate_gsea}
\title{Iterate GSEA}
\usage{
iterate_gsea(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  quantize = list(x = 10, y = 10),
  use_quantiles = list(x = 2, y = 2),
  progressbar = TRUE,
  workers = 1,
  verbose = TRUE
)
}
\arguments{
\item{xmat}{gene x trait matrix.}

\item{ymat}{gene x celltype matrix.}

\item{correction_method}{Multiple-testing correction
method to be passed to \link[stats]{p.adjust}.}

\item{qvalue_thresh}{q-value threshold to use when report
significant results summary.}

\item{quantize}{A named list where the values of "x" and "y" indicate the
number of quantiles to bin the respective \code{xmat} and
\code{ymat} datasets into.}

\item{use_quantiles}{Which quantiles in to use when filtering genes in each column of
\code{xmat} and \code{ymat}.}

\item{progressbar}{logical(1) Enable progress bar
(based on \code{plyr:::progress_text}).
 Enabling the progress bar changes the default value of tasks to
 \code{.Machine$integer.max}, so that progress is reported for
 each element of X.}

\item{workers}{Number (>1) or proportion (<1) of worker cores to use.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} of enrichment results.
}
\description{
Iterate Gene Set Enrichment Analysis (GSEA) across all traits and cell types.
}
\examples{
set.seed(42)

### x matrix (e.g. cell type expression specificity signatures)
xmat <- simulate_matrix(nrow = 200, ncol = 10, col_prefix="celltype_")

### y matrix (e.g. gene-phenotype association signatures)
ymat <- simulate_matrix(nrow = 100, ncol = 50, col_prefix="phenotype_") 

res_gsea <- iterate_gsea(xmat = xmat,
                         ymat = ymat,
                         workers = 1)
}
